# TestJavaProgrammer/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Customer Management API
  description: |
    REST API for managing customer data. Provides complete CRUD operations
    for customers with validation, pagination support, and consistent error handling.
    
    ## Features
    - **CRUD Operations**: Create, Read, Update, and Delete customers
    - **Pagination**: Paginated list endpoints with metadata
    - **Validation**: Input validation with detailed error messages
    - **Error Handling**: Consistent error response format
    
    ## Data Storage
    Customer data is persisted to a local JSON file for simplicity.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Local Development Server
  - url: https://api.example.com
    description: Production Server

tags:
  - name: Customers
    description: Customer management operations

paths:
  /api/customers:
    get:
      tags:
        - Customers
      summary: List all customers
      description: |
        Retrieves a paginated list of all customers. 
        Supports pagination through `page` and `size` query parameters.
      operationId: getAllCustomers
      parameters:
        - name: page
          in: query
          description: Page number (zero-based)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          example: 0
        - name: size
          in: query
          description: Number of records per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 20
      responses:
        '200':
          description: Successfully retrieved paginated list of customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerPageResponse'
              example:
                content:
                  - id: "123e4567-e89b-12d3-a456-426614174000"
                    name: "John Doe"
                    email: "john.doe@example.com"
                  - id: "987fcdeb-51a2-3c4d-b567-890123456789"
                    name: "Jane Smith"
                    email: "jane.smith@example.com"
                page: 0
                size: 20
                totalElements: 2
                totalPages: 1
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Customers
      summary: Create a new customer
      description: |
        Creates a new customer with the provided data.
        A unique ID will be generated automatically if not provided.
      operationId: addCustomer
      requestBody:
        description: Customer data to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
            example:
              name: "John Doe"
              email: "john.doe@example.com"
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
              example:
                id: "123e4567-e89b-12d3-a456-426614174000"
                name: "John Doe"
                email: "john.doe@example.com"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2026-02-01T10:30:00Z"
                status: 400
                error: "Bad Request"
                message: "Validation failed"
                path: "/api/customers"
                errors:
                  - field: "name"
                    message: "Name is required"
                  - field: "email"
                    message: "Email must be valid"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/customers/{id}:
    get:
      tags:
        - Customers
      summary: Get a customer by ID
      description: Retrieves a specific customer by their unique identifier.
      operationId: getCustomer
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
              example:
                id: "123e4567-e89b-12d3-a456-426614174000"
                name: "John Doe"
                email: "john.doe@example.com"
        '404':
          description: Customer not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Customers
      summary: Update a customer
      description: |
        Updates an existing customer with the provided data.
        The customer ID in the path takes precedence over any ID in the request body.
      operationId: updateCustomer
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      requestBody:
        description: Updated customer data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
            example:
              name: "John Updated"
              email: "john.updated@example.com"
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
              example:
                id: "123e4567-e89b-12d3-a456-426614174000"
                name: "John Updated"
                email: "john.updated@example.com"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2026-02-01T10:30:00Z"
                status: 400
                error: "Bad Request"
                message: "Validation failed"
                path: "/api/customers/123e4567-e89b-12d3-a456-426614174000"
                errors:
                  - field: "email"
                    message: "Email must be valid"
        '404':
          description: Customer not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Customers
      summary: Delete a customer
      description: Deletes a customer by their unique identifier.
      operationId: deleteCustomer
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '204':
          description: Customer deleted successfully
        '404':
          description: Customer not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Customer:
      type: object
      description: A customer entity with all fields
      required:
        - id
        - name
        - email
      properties:
        id:
          type: string
          description: Unique identifier for the customer (UUID format)
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          description: Full name of the customer
          minLength: 1
          example: "John Doe"
        email:
          type: string
          format: email
          description: Email address of the customer
          example: "john.doe@example.com"

    CustomerRequest:
      type: object
      description: Request body for creating or updating a customer
      required:
        - name
        - email
      properties:
        id:
          type: string
          description: Optional customer ID (auto-generated if not provided)
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          description: Full name of the customer
          minLength: 1
          example: "John Doe"
        email:
          type: string
          format: email
          description: Email address of the customer
          example: "john.doe@example.com"

    CustomerPageResponse:
      type: object
      description: Paginated response containing customers and pagination metadata
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          description: List of customers for the current page
          items:
            $ref: '#/components/schemas/Customer'
        page:
          type: integer
          description: Current page number (zero-based)
          minimum: 0
          example: 0
        size:
          type: integer
          description: Number of items per page
          minimum: 1
          example: 20
        totalElements:
          type: integer
          format: int64
          description: Total number of customers across all pages
          minimum: 0
          example: 42
        totalPages:
          type: integer
          description: Total number of pages
          minimum: 0
          example: 3

    ErrorResponse:
      type: object
      description: Standard error response structure
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: Time when the error occurred
          example: "2026-02-01T10:30:00Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: HTTP status reason phrase
          example: "Bad Request"
        message:
          type: string
          description: Human-readable error message
          example: "Validation failed"
        path:
          type: string
          description: Request path that caused the error
          example: "/api/customers"
        errors:
          type: array
          description: List of field-level validation errors (if applicable)
          nullable: true
          items:
            $ref: '#/components/schemas/FieldError'

    FieldError:
      type: object
      description: Field-level validation error details
      required:
        - field
        - message
      properties:
        field:
          type: string
          description: Name of the field that failed validation
          example: "email"
        message:
          type: string
          description: Validation error message
          example: "Email must be valid"

  parameters:
    CustomerId:
      name: id
      in: path
      description: Unique identifier of the customer
      required: true
      schema:
        type: string
      example: "123e4567-e89b-12d3-a456-426614174000"